requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
pytest==7.2.1
requests==2.28.2
selenium==4.8.0
undetected-chromedriver==3.4.6

